build DropsUI: Add status signaling to buttons
Changes backcolor of funding available buttons in the main page
UI based on comparison to budget.

Additionally, creates a class to contain sample data for code
building until database can carry the load. This allows focus
on UI logic without burden of interfacing to the data structure
at this point.